requirements.txt
requests==2.32.3
lxml==5.2.1
pandas==2.2.2
xmlschema==3.3.1
reportlab==4.1.0
jupyterlab==4.1.6
python-dotenv==1.0.1 